chore: rename abis -> abi
